requirements.txt
alembic==1.13.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
